Criação da camada service e Controllers de Estado e Cidade

+++ src/main/java/com/github/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,13 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
 	@Transactional
 	@Override
-	public void remover(Estado estado) {
-		estado = buscar(estado.getId());
+	public void remover(Long id) {
+		Estado estado = buscar(id);
+
+		if (estado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+
 		em.remove(estado);
 	}
